pyproject.toml
[tool.poetry]
name = "ollama-x"
version = "0.1.0"
description = "Ollama Proxy Server with Extensions"
authors = ["Yury Sokov (Yurzs) <me@yurzs.dev>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = ">=2.8.2,<3"
uvicorn = ">=0.30,<1"
fastapi = ">=0.112,<1"
pydantic-mongo-document = {version = ">=0.2", extras = ["asyncio"]}
aiohttp = "^3.10.3"
pydantic-app-config = {git = "https://github.com/Yurzs/pydantic-app-config.git"}
apscheduler = "^3.10.4"
langfuse = "^2.43.3"
email-validator = "^2.2.0"
sentry-sdk = {extras = ["fastapi"], version = "^2.13.0"}
openapi-cli = "^0.2"
sse-starlette = "^2.1.3"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.8.0"
autoflake = "^2.3.1"
plumbum = "^1.8.3"
ruff = "^0.6.2"
openapi-cli = {extras = ["full"], version = "^0.2"}
ollama-x-client = "^0.1.0"

[tool.poetry.scripts]
ollama-x = "openapi_cli.cli:main"
ollama-x-scheduler = "ollama_x.scheduler:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100

[tool.black]
line_length = 100
target_version = ["py312"]

[tool.pylint]
target_versions = ["py312"]
max-line-length = 100
profile = "black"

[tool.autoflake]
remove_all_unused_imports = true
remove_unused_variables = true
in-place = true
recursive = true

[tool.ruff]
line-length = 100

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[toll.ruff.format]
docstring-code-format = true
quote-style = "double"
indent-style = "space"
